# jarkata/spring-data-jpa/src/main/java/org/howe/examples/jpa/entity/Preference.java
package org.howe.examples.jpa.entity;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Preference {
    private String hobby;
    private int experience;
}